added numberPicker to cameraView. needs default string value
