+++ app/src/main/java/com/example/fair_share/takeCameraView.java
@@ -8,7 +8,11 @@ import android.widget.ImageButton;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.FragmentTransaction;
 
+import com.shawnlin.numberpicker.NumberPicker;
+
 public class takeCameraView extends AppCompatActivity {
+
+    private int selectedValue = 4; // Default value
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -20,6 +24,12 @@ public class takeCameraView extends AppCompatActivity {
 
         //initialize buttons
         ImageButton startCapture_button = findViewById(R.id.startCapture_button);
+
+        // Initialize onClick methods
+        startCapture_button.setOnClickListener(this::handleButtonClick);
+
+        // Initialize NumberPicker
+        NumberPicker();
 
         //initialize onClick methods
         startCapture_button.setOnClickListener(this::handleButtonClick);
@@ -33,12 +43,29 @@ public class takeCameraView extends AppCompatActivity {
         transaction.commit();
     }
 
+    public void NumberPicker(){
+        NumberPicker numberPicker = findViewById(R.id.number_picker);
+
+        numberPicker.setMinValue(2);
+        numberPicker.setMaxValue(10);
+        numberPicker.setValue(4);
+
+        //values change while scrolling
+        numberPicker.setOnValueChangedListener(new NumberPicker.OnValueChangeListener() {
+            @Override
+            public void onValueChange(NumberPicker picker, int oldVal, int newVal) {
+                selectedValue = newVal;
+            }
+        });
+    }
+
     //ALL ONCLICK EVENTS HERE
     private void handleButtonClick(View view) {
         int id = view.getId();
 
         if (id == R.id.startCapture_button) {
             Intent intent = new Intent(this, listReceipt.class);
+            intent.putExtra("selectedValue", selectedValue);
             startActivity(intent);
         }
     }
