+++ app/src/main/res/layout/activity_list_reciept.xml
@@ -29,27 +29,7 @@
         app:layout_constraintVertical_bias="0.023"
         app:srcCompat="@drawable/logotext" />
 
-    <com.shawnlin.numberpicker.NumberPicker
-        android:id="@+id/number_picker"
-        android:layout_width="90dp"
-        android:layout_height="109dp"
-        android:layout_centerInParent="true"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.498"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.379"
-        app:np_dividerColor="@android:color/black"
-        app:np_height="180dp"
-        app:np_max="10"
-        app:np_min="0"
-        app:np_selectedTextColor="@android:color/black"
-        app:np_selectedTextSize="18sp"
-        app:np_textColor="@android:color/black"
-        app:np_textSize="14sp"
-        app:np_value="5"
-        app:np_width="64dp" />
+
 
     <LinearLayout
         android:id="@+id/buttonContainer"
