+++ app/src/main/res/layout/activity_take_camera_view.xml
@@ -64,7 +64,29 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.32"
         app:srcCompat="@drawable/cameraframe"
-        app:tint="#6B674FA2" />
+        app:tint="#5B000000" />
+
+    <com.shawnlin.numberpicker.NumberPicker
+        android:id="@+id/number_picker"
+        android:layout_width="90dp"
+        android:layout_height="109dp"
+        android:layout_centerInParent="true"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.498"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.974"
+        app:np_dividerColor="@color/button_color"
+        app:np_height="180dp"
+        app:np_max="10"
+        app:np_min="0"
+        app:np_selectedTextColor="@color/white"
+        app:np_selectedTextSize="18sp"
+        app:np_textColor="#BBB6B6"
+        app:np_textSize="14sp"
+        app:np_value="5"
+        app:np_width="64dp" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
